pyproject.toml
[project]
name = "chrono-wallpaper"
version = "1.0.0"
description = "Time-based wallpaper transitions for Wayland compositors"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Tom Doerr", email = "tom@doerr.dev"}
]
keywords = ["wallpaper", "wayland", "swaybg", "hyprland", "transition"]

dependencies = [
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "click>=8.1.0",
    "psutil>=5.9.0",
    "tomli>=2.0.0; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
chrono-wallpaper = "chrono_wallpaper.cli.main:cli"

[project.urls]
Homepage = "https://github.com/tom-doerr/chrono-wallpaper"
Repository = "https://github.com/tom-doerr/chrono-wallpaper"
Issues = "https://github.com/tom-doerr/chrono-wallpaper/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/chrono_wallpaper"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=chrono_wallpaper"
